+++ src/main/java/br/com/etyllica/animation/script/text/DialogScript.java
@@ -1,4 +1,4 @@
-package br.com.etyllica.core.animation.script.text;
+package br.com.etyllica.animation.script.text;
 
 import br.com.etyllica.core.animation.script.AnimationScript;
 import br.com.etyllica.layer.TextLayer;
