+++ src/main/java/br/com/etyllica/core/InnerCore.java
@@ -14,6 +14,7 @@ import br.com.etyllica.core.animation.script.SingleIntervalAnimation;
 import br.com.etyllica.core.context.Application;
 import br.com.etyllica.core.context.Context;
 import br.com.etyllica.core.context.UpdateIntervalListener;
+import br.com.etyllica.core.effect.GlobalEffect;
 import br.com.etyllica.core.event.GUIEvent;
 import br.com.etyllica.core.event.KeyEventListener;
 import br.com.etyllica.core.event.KeyEvent;
@@ -28,7 +29,6 @@ import br.com.etyllica.core.i18n.LanguageChangerListener;
 import br.com.etyllica.core.i18n.LanguageHandler;
 import br.com.etyllica.core.input.keyboard.Keyboard;
 import br.com.etyllica.core.input.mouse.Mouse;
-import br.com.etyllica.effects.GlobalEffect;
 import br.com.etyllica.gui.View;
 import br.com.etyllica.gui.Window;
 import br.com.etyllica.theme.Theme;
