+++ src/main/java/br/com/etyllica/core/effect/particle/Particle.java
@@ -1,0 +1,12 @@
+package br.com.etyllica.core.effect.particle;
+
+import br.com.etyllica.core.Updatable;
+import br.com.etyllica.layer.Layer;
+
+public abstract class Particle extends Layer implements Updatable {
+	
+	public Particle(int x, int y) {
+		super(x,y);
+	}
+	
+}
