+++ src/main/java/br/com/etyllica/effects/GenericFullScreenEffect.java
@@ -4,6 +4,7 @@ import java.awt.Color;
 
 import br.com.etyllica.core.Configuration;
 import br.com.etyllica.core.animation.script.OpacityAnimation;
+import br.com.etyllica.core.effect.GlobalEffect;
 import br.com.etyllica.core.graphics.Graphic;
 import br.com.etyllica.core.i18n.DefaultDictionary;
