+++ src/main/java/examples/etyllica/tutorial20/AnimatedDialog.java
@@ -1,7 +1,7 @@
 package examples.etyllica.tutorial20;
 
+import br.com.etyllica.animation.script.text.DialogScript;
 import br.com.etyllica.core.animation.AnimationHandler;
-import br.com.etyllica.core.animation.script.text.DialogScript;
 import br.com.etyllica.core.context.Application;
 import br.com.etyllica.core.event.GUIEvent;
 import br.com.etyllica.core.event.KeyEvent;
