+++ src/main/java/examples/fx/application/LightningApplication.java
@@ -3,11 +3,11 @@ package examples.fx.application;
 import java.awt.Color;
 
 import br.com.etyllica.core.context.Application;
+import br.com.etyllica.core.effect.Effect;
 import br.com.etyllica.core.event.GUIEvent;
 import br.com.etyllica.core.event.KeyEvent;
 import br.com.etyllica.core.event.PointerEvent;
 import br.com.etyllica.core.graphics.Graphic;
-import br.com.etyllica.effects.Effect;
 
 public class LightningApplication extends Application{
